Phase B-7 (FINAL): kx-mote lib.rs split (Rule 3, behavior-preserving; load-bearing trait surface)
FINAL Phase B refactor per Rule 3. **Rule 1 max-strict surface** — kx-mote
is the narrow waist of every downstream kortecx crate; the load-bearing
trait surface for journal, projection, executor, scheduler, runtime.
Pure move; zero functional change.

Before: src/lib.rs 1332 lines (5 hash newtypes + 6 string newtypes + 3
public enums + 3 edge types + MoteDef + canonical_config + derive_mote_id
+ Mote + 5-piece attempt state machine + 3 D37 topology primitives +
MoteGraph + 17 inline tests).

After: thin src/lib.rs 95 lines + 10 sibling modules organized by
concern (matches the worked example in 07-engineering-discipline.md
§Rule 3, expanded for the full public surface):

- id.rs (Hash32 alias + MoteId + MoteDefHash + InputDataId + LogicRef +
  PromptTemplateHash + their Debug/Display impls)
- strings.rs (ModelId + ToolName + ToolVersion + ConfigKey + ConfigVal +
  GraphPosition)
- ndclass.rs (NdClass enum + impl)
- effect.rs (EffectPattern enum)
- edge.rs (EdgeKind + EdgeMeta + ParentRef + impls)
- def.rs (MOTE_DEF_SCHEMA_VERSION + MoteDef + impl + canonical_config +
  derive_mote_id — the identity-derivation core)
- mote.rs (Mote + impl Mote)
- attempt.rs (AttemptState + IllegalTransition + transition +
  is_legal_transition + ALL_ATTEMPT_STATES — the lifecycle state machine)
- topology.rs (RoleId + ChildDescriptor + TopologyDecision + impl —
  D37 Seam A enforcement primitives)
- graph.rs (MoteGraph + impl)
- tests.rs (17 inline unit tests; #[cfg(test)] from lib.rs)

Re-exports preserve every previously-public path. No downstream crate's
use statements needed editing.

Lesson from B-6 applied preemptively: ran `RUSTDOCFLAGS=-D warnings
cargo doc -p kx-mote --no-deps` BEFORE push to catch intra-doc-link
breakage. ~16 link references updated to use `crate::` prefix where they
cross sibling-module boundaries (e.g., `[\`MoteDef\`]` →
`[\`crate::MoteDef\`]` when referenced from id.rs / strings.rs / topology.rs).

Local gates green on Apple Silicon (--all-features):
- cargo build -p kx-mote: clean
- cargo test --workspace --all-features: 645 + 8 (IDENTICAL)
- cargo test -p kx-mote: 17 + 3 + 16 + 13 + 1 + 8 = 58 (IDENTICAL)
- cargo clippy -p kx-mote --all-targets -- -D warnings: clean
- cargo fmt --all -- --check: clean
- RUSTDOCFLAGS=-D warnings cargo doc -p kx-mote --no-deps: clean

**PHASE B COMPLETE after this PR merges.** All 7 monolithic-lib.rs crates
are now split per Rule 3:
  B-1 kx-model-validator (1187 → 95)
  B-2 kx-tool-registry (1504 → 105)
  B-3 kx-warrant (1254 → 100)
  B-4 kx-context-assembler (949 → 105)
  B-5 kx-capability (1304 → 110) + dual-repo drift catch-up
  B-6 kx-projection (1305 → 95)
  B-7 kx-mote (1332 → 95) ← THIS PR

Next: PR 10 (kx-scheduler / P1.10) — first feature ticket governed by
all four Golden Rules from day one.
